Changed layout manager to be recursive (fixes #12).
Now layout is achieved by starting with the reference component and laying everything out based on its links, in a recursive fashion. When a double link is found, i.e. a component is linked to another component via two different paths, then the link that would form the double path is drawn as a straight line between the two nodes (usually a non-physical link).

Automatic resolution of doubly-linked components is difficult and will be left to a future version of Optivis. For now, an ugly straight line will hopefully be enough to show the user they need to be careful.